# internal/server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: cueBreaker API
  description: |
    JSON API for splitting single-file FLAC albums using their CUE sheets.

    Every path below is served by the cueBreaker backend alongside the embedded
    SPA. This spec is hand-written and kept honest by a drift test that compares
    it against the routes the server actually registers.
  version: "1"
servers:
  - url: /
    description: The cueBreaker backend itself
...
paths:
...
  /api/preview:
    post:
      tags: [Library]
      summary: Parse a CUE sheet and describe its album
      description: |
        Parses the named CUE sheet and returns its album metadata and tracks,
        plus the cover, split-status and source-duration data the waveform
        needs.
      operationId: preview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PathRequest"
      responses:
        "200":
          description: The parsed album.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preview"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
...
components:
  schemas:
...
    Track:
      type: object
      description: A single CUE track entry.
      required: [number, title, performer, index, start_seconds]
      properties:
        number:
          type: integer
        title:
          type: string
        performer:
          type: string
        index:
          type: string
          description: The raw CUE INDEX timecode, `mm:ss:ff`.
        start_seconds:
          type: number
          format: double
          description: The track's start offset, in seconds.
    Album:
      type: object
      description: A parsed CUE sheet's album metadata and tracks, in CUE order.
      required: [performer, title, file, genre, date, tracks]
      properties:
        performer:
          type: string
        title:
          type: string
        file:
          type: string
          description: The source file named by the CUE sheet's FILE line.
        genre:
          type: string
        date:
          type: string
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
    Preview:
      allOf:
        - $ref: "#/components/schemas/Album"
        - type: object
          required: [has_cover, split_done, output_tracks, total_seconds]
          properties:
            has_cover:
              type: boolean
            cover_name:
              type: string
              description: The cover's file name. Omitted when there is no cover.
            split_done:
              type: boolean
            output_tracks:
              type: integer
            total_seconds:
              type: number
              format: double
              description: |
                The source file's duration. `0` when it could not be read.
    PathRequest:
      type: object
      description: The shared body of the preview and split endpoints.
      required: [path, cue_file]
      properties:
        path:
          type: string
          description: |
            The album directory's path, relative to the input root. Absolute
            paths and any `..` that escapes the root are rejected with a 403.
        cue_file:
          type: string
          description: The CUE sheet's file name within that directory.
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
    Error:
      description: The request failed; the body carries the reason.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"